requirements.txt
tensorflow==2.13.0
opencv-python==4.8.0.74
mediapipe==0.10.3
numpy==1.24.3
scipy==1.11.1
scikit-learn==1.3.0
Pillow==10.0.0
PyYAML==6.0
loguru==0.7.0
adafruit-circuitpython-ssd1306==2.12.14
adafruit-blinka==8.20.0
RPi.GPIO==0.7.1
board==1.0
busio==5.3.4
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
tqdm==4.65.0
click==8.1.6
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
isort==5.12.0
